requirements.txt
Flask
gunicorn
Pillow
opencv-python
numpy